[fix] Remove all implicit type conversions.

+++ src/Cursor.cpp
@@ -20,27 +20,27 @@ SDL_FRect* Cursor::get_rectangle() {
       return &rectangle;
 }
 
-float Cursor::get_r() const {
+Uint8 Cursor::get_r() const {
       return color.r;
 }
 
-float Cursor::get_g() const {
+Uint8 Cursor::get_g() const {
       return color.g;
 }
-float Cursor::get_b() const {
+Uint8 Cursor::get_b() const {
       return color.b;
 }
-float Cursor::get_a() const {
+Uint8 Cursor::get_a() const {
       return color.a;
 }
 
 void Cursor::set_column(int x) {
-      rectangle.x = x * rectangle.w;
+      rectangle.x = static_cast<float>(x) * rectangle.w;
       column      = x;
 }
 
 void Cursor::set_row(int y) {
-      rectangle.y = y * rectangle.h;
+      rectangle.y = static_cast<float>(y) * rectangle.h;
       row         = y;
 }
